+++ src/line.fs
@@ -1,6 +1,6 @@
 const FRAGMENT_SHADER = `
 precision mediump float;
-uniform vec4 color;
+varying vec4 color;
 void main() {
   gl_FragColor = color;
 }`;